MVP for client side implementation (for experimentation) of shortest path planner

+++ jNoRightTurns/norightturns/src/main/java/com/ken/norightturns/util/IDGenerator.java
@@ -1,9 +1,9 @@
 package com.ken.norightturns.util;
 
 public class IDGenerator {
-	private long idAccum = 0;
-	public long get() {
-		long id = idAccum;
+	private int idAccum = 0;
+	public int get() {
+		int id = idAccum;
 		idAccum++;
 		return id;
 	}
